feat(components): Introduce Base, Heater, and Temperature classes with NTC sensor implementation
- Added Base class as an abstract interface for components with setup and loop methods.
- Implemented Heater class inheriting from Base, defining target temperature methods.
- Created TemperatureSensor class extending Base, with a method to get temperature.
- Developed NTC_Sensor_Temperature class implementing TemperatureSensor, including temperature reading and task management.
- Added HeaterSSR class for controlling heating elements with PID control and temperature monitoring.
- Refactored main and module files to integrate new components and improve temperature handling.

+++ src/Settings.h
@@ -49,10 +49,10 @@ class Settings {
 
         String wifiSsid = "";
         String wifiPassword = "";
-        int kp = 2;
-        int ki = 5;
-        int kd = 1;
-        int pOn = 1000;
+        float kp = 2;
+        float ki = 5;
+        float kd = 1;
+        float pOn = 1000;
         int time = 1000;
         int volumeLiters = 70;
         int powerWatts = 3200;
